README: Add instructions for uxnemu on Windows
When compiled under the MSYS2 environment, the executible that is
produced expects the SDL2 library to be available at runtime. When
running outside MSYS2, that may not be the case, and the emulator may
crash on startup with a message about missing SDL2.dll, or sometimes
without a message at all.

This patch adds some guidance for this scenario by guiding the reader
to download the precompiled SDL binaries for Windows and make them
available for the emulator's use.