Reuse old artifacts if no solar build computers are reachable.
Non-Linux builds are run on computers powered by solar panels in one
geographical location, so they are not available 100% of the time. When
they are not available, the build script will reuse the previous build
artifacts for Windows and macOS, as well as create an empty file at
https://rabbits.srht.site/uxn/needs-solar-build so that this situation
can be detected. When both build computers are available, the
needs-solar-build file will not be present in the Sourcehut upload, so
that URL will return a 404 status.